% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.se.r
\name{bs.se}
\alias{bs.se}
\title{Performs a parametric bootstrapping standard error approximation}
\usage{
bs.se(output, B = 100)
}
\arguments{
\item{output}{output of \code{\link{StagedChoiceSplineMix}}}

\item{B}{number of bootstrap samples (def:100)}
}
\description{
The function performs a parametric bootstrapping standard error approximation for mixtures of two-stage logistic regressions.
}
\examples{
## parametric bootstrapping to calculate standard error:
## use best output from StagedChoiceSplineMix exampl
#out.se<-bs.se(output=out$best, B=100)
#out.se$betab.se
#out.se$betaw.se
#out.se$lambda.se
}
\seealso{
\code{\link{StagedChoiceSplineMix}} \cr
\code{\link[mixtools]{boot.se}}
}

