% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.init.r
\name{gen.init}
\alias{gen.init}
\title{Generates initial values for mixtures of logistic regressions}
\usage{
gen.init(y, x, k, er)
}
\arguments{
\item{y}{See \code{\link{StagedChoiceSplineMix}} for details.}

\item{x}{See \code{\link{StagedChoiceSplineMix}} for details.}

\item{k}{See \code{\link{StagedChoiceSplineMix}} for details.}

\item{er}{The total number of errors. See \code{\link{StagedChoiceSplineMix}} for details.}
}
\description{
A sub-function of \code{\link{StagedChoiceSplineMix}}. This function generates initial values for mixtures of logistic regressions. This function is used if starting points for parameters are not specified by the user or when the EM algorithm needs to be initialized due to errors.
}
\seealso{
\code{\link{StagedChoiceSplineMix}} \cr
"mixtools" package version 1.0.3
}

