% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.knot.r
\name{move.knot}
\alias{move.knot}
\title{Generates a new set of knots for the following iteration}
\usage{
move.knot(num.knot, sp.knots, k)
}
\arguments{
\item{num.knot}{See \code{\link{StagedChoiceSplineMix}} for details.}

\item{sp.knots}{See \code{\link{StagedChoiceSplineMix}} for details.}

\item{k}{See \code{\link{StagedChoiceSplineMix}} for details.}
}
\description{
A sub-function of \code{\link{StagedChoiceSplineMix}}. This function generates a new set of knots for the following iteration. Please refer to Bruch et al. (in press) for the precise rule used.
}
\references{
Bruch, E., F. Feinberg, K. Lee (in press), "Detecting Cupid's Vector: Extracting Decision Rules from Online Dating Activity Data," \emph{Proceedings of the National Academy of Sciences}.
}
\seealso{
\code{\link{StagedChoiceSplineMix}}
}

